Update to the latest CSWin32 (#9300)
Take the latest drop of CSWin32.

There were a number of changes, but the key ones were:

- VARIANT moved to another namespace
- CsWin32 wrapper methods that take strings no longer have unusued/reserved parameters
- HGLOBAL is now used in signatures

A few other things in this change:

- Added start-vs.cmd to the solution folder
- Pinned created Arrays in VARIANT to avoid a potential memory corruption
- Removed VARIANT IRecord related code that was ultimately going to throw

+++ src/System.Windows.Forms/src/System/Windows/Forms/AxHost.cs
@@ -1125,7 +1125,7 @@ public abstract unsafe partial class AxHost : Control, ISupportInitialize, ICust
         ((IOleControlSite.Interface)_oleSite).TransformCoords(
             (POINTL*)&phm,
             &pcont,
-            XFORMCOORDS.XFORMCOORDS_SIZE | XFORMCOORDS.XFORMCOORDS_HIMETRICTOCONTAINER);
+            (uint)(XFORMCOORDS.XFORMCOORDS_SIZE | XFORMCOORDS.XFORMCOORDS_HIMETRICTOCONTAINER));
 
         sz.Width = (int)pcont.X;
         sz.Height = (int)pcont.Y;
@@ -1138,7 +1138,7 @@ public abstract unsafe partial class AxHost : Control, ISupportInitialize, ICust
         ((IOleControlSite.Interface)_oleSite).TransformCoords(
             (POINTL*)&phm,
             &pcont,
-            XFORMCOORDS.XFORMCOORDS_SIZE | XFORMCOORDS.XFORMCOORDS_CONTAINERTOHIMETRIC);
+            (uint)(XFORMCOORDS.XFORMCOORDS_SIZE | XFORMCOORDS.XFORMCOORDS_CONTAINERTOHIMETRIC));
 
         sz.Width = phm.X;
         sz.Height = phm.Y;
